Commit migrated from https://github.com/dotnet/coreclr/commit/b2c397b649181f793d6c29f1de78d0aee0a52596

+++ src/coreclr/src/jit/utils.cpp
@@ -2153,9 +2153,9 @@ const SignedMagic<int32_t>* TryGetSignedMagic(int32_t divisor)
     static const SignedMagic<int32_t> table[]{
         {0x55555556, 0}, // 3
         {},
-        {0x66666667, 1}, // 5
-        {0x2aaaaaab, 0}, // 6
-        {0x92492493, 2}, // 7
+        {0x66666667, 1},          // 5
+        {0x2aaaaaab, 0},          // 6
+        {(int32_t)0x92492493, 2}, // 7
         {},
         {0x38e38e39, 1}, // 9
         {0x66666667, 2}, // 10
